+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # In Memory Cache Class wrapper
 
+## 2.0.0
+ - feat: Added support for redis cache
+ - breaking changes: Initialization - getCacheInstance() function is now asynchronous
+ - feat: Added a destroy functionality for the cache helper instance.
+
+## 1.0.1
+ - fix: Class initialization - options parameter is now mandatory
+
 ## 1.0.1
  - fix: Class initialization - options parameter is now mandatory